construcao dos builders

+++ src/main/java/challenge/Estacionamento.java
@@ -30,7 +30,7 @@ public class Estacionamento {
     }
 
     // carros estacionados
-    public Integer carrosEstacionados() {
+    public long carrosEstacionados() {
         return carros.size();
     }
